[FIX] PID started at record and source

+++ src/cta.c
@@ -19,36 +19,36 @@ static unsigned long long last_pid      = 0;
 long cta_state_machine(aSubRecord* prec) {
 
 	// Inputs
-	unsigned short start      = *(unsigned short*)     prec->a;
-	unsigned short stop       = *(unsigned short*)     prec->b;
-	unsigned short index      = *(unsigned short*)     prec->c;
-	unsigned short running    = *(unsigned short*)     prec->d;
-	unsigned long long pid    = *(unsigned long long*) prec->e;
-	unsigned short started_at = *(unsigned short*)     prec->f;
-	unsigned short missed_pid = *(unsigned short*)     prec->g;
-	unsigned short length     = *(unsigned short*)     prec->h;
-	unsigned short cycles     = *(unsigned short*)     prec->i;
-	unsigned short cfgMod     = *(unsigned short*)     prec->j;
-	unsigned short cfgModDiv  = *(unsigned short*)     prec->k;
-	unsigned short cfgModOff  = *(unsigned short*)     prec->l;
+	unsigned short start          = *(unsigned short*)     prec->a;
+	unsigned short stop           = *(unsigned short*)     prec->b;
+	unsigned short index          = *(unsigned short*)     prec->c;
+	unsigned short running        = *(unsigned short*)     prec->d;
+	unsigned long long pid        = *(unsigned long long*) prec->e;
+	unsigned long long started_at = *(unsigned long long*) prec->f;
+	unsigned short missed_pid     = *(unsigned short*)     prec->g;
+	unsigned short length         = *(unsigned short*)     prec->h;
+	unsigned short cycles         = *(unsigned short*)     prec->i;
+	unsigned short cfgMod         = *(unsigned short*)     prec->j;
+	unsigned short cfgModDiv      = *(unsigned short*)     prec->k;
+	unsigned short cfgModOff      = *(unsigned short*)     prec->l;
       
         // Outputs
-	unsigned short *out_start;
-	unsigned short *out_stop;
-	unsigned short *out_index;
-	unsigned short *out_running;
-	unsigned short *out_started_at;
-	unsigned short *out_missed_pid;
-	unsigned short *out_load_seq;
-	unsigned short *out_enable_evt;
-	out_start      = (unsigned short*) prec->vala;
-	out_stop       = (unsigned short*) prec->valb;
-	out_index      = (unsigned short*) prec->valc;
-	out_running    = (unsigned short*) prec->vald;
-	out_started_at = (unsigned short*) prec->vale;
-	out_missed_pid = (unsigned short*) prec->valf;
-	out_load_seq   = (unsigned short*) prec->valg;
-	out_enable_evt = (unsigned short*) prec->valh;
+	unsigned short     *out_start;
+	unsigned short     *out_stop;
+	unsigned short     *out_index;
+	unsigned short     *out_running;
+	unsigned long long *out_started_at;
+	unsigned short     *out_missed_pid;
+	unsigned short     *out_load_seq;
+	unsigned short     *out_enable_evt;
+	out_start      = (unsigned short*)     prec->vala;
+	out_stop       = (unsigned short*)     prec->valb;
+	out_index      = (unsigned short*)     prec->valc;
+	out_running    = (unsigned short*)     prec->vald;
+	out_started_at = (unsigned long long*) prec->vale;
+	out_missed_pid = (unsigned short*)     prec->valf;
+	out_load_seq   = (unsigned short*)     prec->valg;
+	out_enable_evt = (unsigned short*)     prec->valh;
 
 	// State update
 	if(stop)         state = STOPPED;
